refactor(openclaw): split reply listener facade
