+++ src/openclaw/reply-listener-signature.ts
@@ -1,0 +1,5 @@
+import type { OpenClawConfig } from "./types"
+
+export function getReplyListenerRuntimeSignature(config: Pick<OpenClawConfig, "replyListener"> | null): string {
+  return JSON.stringify(config?.replyListener ?? null)
+}
